moved context folder, hooks to a higher level
