+++ src/components/profile/PostCard.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useModal } from "../hooks/useModal";
+import { useModal } from "../../hooks/useModal";
 import CommentMain from "../main/maincomponents/comments/CommentMain";
 import { BsFillChatFill, BsHeart, BsHeartFill } from "react-icons/bs";
