+++ src/hooks/useCreateComment.jsx
@@ -1,7 +1,6 @@
 import { useContext } from "react";
-import { CreateNewComment } from "../../services/comments";
+import { CreateNewComment } from "../services/comments";
 import { AuthContext } from "../contexts/AuthContext";
-import { getPostsContext } from "../contexts/GetPostsContext";
 
 export function useCreateComment(idPost) {
   const { getToken } = useContext(AuthContext);
